Rename LoginActivityModule to MainActivityModule
