+++ app/src/main/java/com/edvinas/balkaitis/party/MainActivityModule.kt
@@ -1,4 +1,4 @@
-package com.edvinas.balkaitis.party.login
+package com.edvinas.balkaitis.party
 
 import com.edvinas.balkaitis.party.login.fragment.LoginFragment
 import com.edvinas.balkaitis.party.login.fragment.LoginModule
@@ -7,7 +7,7 @@ import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
 @Module
-abstract class LoginActivityModule {
+abstract class MainActivityModule {
     @FragmentScope
     @ContributesAndroidInjector(modules = [LoginModule::class])
     abstract fun contributeLoginFragment(): LoginFragment
