+++ app/src/main/java/com/edvinas/balkaitis/party/app/AppContributorsModule.kt
@@ -1,7 +1,7 @@
 package com.edvinas.balkaitis.party.app
 
 import com.edvinas.balkaitis.party.MainActivity
-import com.edvinas.balkaitis.party.login.LoginActivityModule
+import com.edvinas.balkaitis.party.MainActivityModule
 import com.edvinas.balkaitis.party.utils.scopes.ActivityScope
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
@@ -9,6 +9,6 @@ import dagger.android.ContributesAndroidInjector
 @Module
 abstract class AppContributorsModule {
     @ActivityScope
-    @ContributesAndroidInjector(modules = [LoginActivityModule::class])
+    @ContributesAndroidInjector(modules = [MainActivityModule::class])
     abstract fun contributePostsActivity(): MainActivity
 }
